==> 02 preprocessing.R <==
load(cards,file='myfolder/03_data/02_aggregated_data/cards_raw.Rdata')
load(questions,file='myfolder/03_data/02_aggregated_data/questions_raw.Rdata')
save(screensize,file='myfolder/03_data/02_aggregated_data/screensize_raw.Rdata')
save(squares,file='myfolder/03_data/02_aggregated_data/squares_raw.Rdata')
